Upgrade Scintilla to 4.1.0 (2018-07-08 7061:d744568dfaa5).

+++ scintilla/lexlib/PropSetSimple.h
@@ -16,11 +16,11 @@ class PropSetSimple {
 public:
 	PropSetSimple();
 	virtual ~PropSetSimple();
-	void Set(const char *key, const char *val, int lenKey=-1, int lenVal=-1);
+	void Set(const char *key, const char *val, size_t lenKey = -1, size_t lenVal = -1);
 	void SetMultiple(const char *);
 	const char *Get(const char *key) const;
 	int GetExpanded(const char *key, char *result) const;
-	int GetInt(const char *key, int defaultValue=0) const;
+	int GetInt(const char *key, int defaultValue = 0) const;
 };
 
 }
